requirements.txt
flask
flask-socketio
python-engineio<4
python-socketio<6
opencv-python
mediapipe
numpy
pydantic
pyyaml